Move ParserC into its own module.
